- fix compile error
https://devzone.nordicsemi.com/question/6950/upgrading-to-sdk-520-breaks
-build/

+++ utility/nrf51822/s110/nrf_svc.h
@@ -16,7 +16,7 @@
   { \
     __asm( \
         "svc %0\n" \
-        "bx r14" : : "I" (number) : "r0" \
+        "bx r14" : : "I" ((uint16_t)number) : "r0" \
     ); \
   }
 #elif defined (__ICCARM__)
@@ -25,7 +25,7 @@
 PRAGMA(swi_number = number) \
  __swi return_type signature;
 #else
-#define SVCALL(number, return_type, signature) return_type signature  
+#define SVCALL(number, return_type, signature) return_type signature
 #endif
 #endif  // SVCALL
